Refactor location search input into separate component
so that we can reuse that functionality in multi-location selection, while continuing to offer comparable functionality in single-location mode.
